Add functionalities to Plateau entity
To allow the new functionalities a new entity has been created to
represent a position. Refactored the Plateau and Rover entities to use
it.

The idea with this functionalities is to have localized the rovers in
the plateau and can check if the next position of the rover is already
occupied.